feat: Added FAQ page, Header component
