+++ text2sql-frontend/index.html
@@ -35,8 +35,8 @@
     <meta name="twitter:image:alt" content="DataLine typed on a blurred background">
 
   </head>
-  <body class="dark:bg-gray-900">
-    <div id="root" class=""></div>
+  <body class="dark:bg-gray-900 h-full">
+    <div id="root" class="h-full"></div>
     <script type="module" src="./src/main.tsx"></script>
     <script
       type="module"
