+++ text2sql-frontend/src/components/Landing/Landing.tsx
@@ -1,12 +1,10 @@
-import { SVGProps, useState } from "react";
-import { Dialog } from "@headlessui/react";
-import { XMarkIcon } from "@heroicons/react/24/outline";
-import { BackgroundLayout } from "../Layouts/BackgroundLayout";
-import logo from "../../assets/images/logo_xl.png";
-import logomd from "../../assets/images/logo_md.png";
+import { SVGProps } from "react";
+
+import { BackgroundLayout } from "@components/Layouts/BackgroundLayout";
+
 import headline_image from "@/assets/images/headline_image.png";
 import edit_query_gif from "@/assets/images/edit_query.gif";
-import connection_creator_image from "@/assets/images/connection_creator.png";
+import data_security_image from "@/assets/images/data_security.png";
 import ReactGA from "react-ga4";
 
 import "./Landing.css";
@@ -15,13 +13,8 @@ import SquareBackground from "@components/Layouts/SquareBackground";
 import SquareBackgroundInverted from "@components/Layouts/SquareBackgroundInverted";
 import BackgroundHighlight from "@components/Layouts/BackgroundHighlight";
 import FeatureComparisonTable from "./FeatureComparisonTable";
+import Header from "@components/Layouts/Header";
 
-const navigation: any[] = [
-  // { name: "Product", href: "#" },
-  // { name: "Features", href: "#" },
-  // { name: "Marketplace", href: "#" },
-  // { name: "Company", href: "#" },
-];
 
 const socials = [
   {
@@ -69,7 +62,7 @@ const CohortData = {
     comparison: {
       title: "How do we compare?",
       description:
-        "This is probably the 10th AI tool you see today, so let's cut to the chase.",
+        "This is probably the 10th AI tool you see today, so we'll cut to the chase. For more info, <a style=\"text-decoration-line: underline;\" href=\"/faq\">our FAQ page</a>",
     },
     cta_title: "Stay up to date.",
     cta_description: "Promise we won't annoy you, ever.",
@@ -86,111 +79,9 @@ if (process.env.NODE_ENV !== "local") {
 }
 
 export const Landing = () => {
-  const [mobileMenuOpen, setMobileMenuOpen] = useState(false);
-
   return (
     <div className="bg-gray-900 overflow-hidden border-r border-white/5">
-      <header className="absolute inset-x-0 top-0 z-50">
-        <nav
-          className="flex items-center justify-between p-6 lg:px-8"
-          aria-label="Global"
-        >
-          <div className="flex lg:flex-1">
-            <a href="#" className="-m-1.5 p-1.5 flex">
-              <span className="sr-only">DataLine</span>
-              <img
-                className="h-6 w-auto mt-1.5"
-                src={logo}
-                alt="DataLine logo made"
-              />
-              <h1 className="ml-3 text-2xl font-semibold tracking-tight text-center text-white sm:text-3xl">
-                DataLine
-              </h1>
-            </a>
-          </div>
-
-          {/* Mobile menu button */}
-          {/* <div className="flex lg:hidden">
-            <button
-              type="button"
-              className="-m-2.5 inline-flex items-center justify-center rounded-md p-2.5 text-gray-400"
-              onClick={() => setMobileMenuOpen(true)}
-            >
-              <span className="sr-only">Open main menu</span>
-              <Bars3Icon className="h-6 w-6" aria-hidden="true" />
-            </button>
-          </div> */}
-
-          {/* <div className="hidden lg:flex lg:gap-x-12">
-            {navigation.map((item) => (
-              <a
-                key={item.name}
-                href={item.href}
-                className="text-sm font-semibold leading-6 text-white"
-              >
-                {item.name}
-              </a>
-            ))}
-          </div> */}
-
-          <div className="hidden lg:flex lg:flex-1 lg:justify-end">
-            {/* <a
-              href="/beta-signup"
-              className="text-sm font-semibold leading-6 text-white"
-            >
-              Sign Up <span aria-hidden="true">&rarr;</span>
-            </a> */}
-          </div>
-        </nav>
-
-        <Dialog
-          as="div"
-          className="lg:hidden"
-          open={mobileMenuOpen}
-          onClose={setMobileMenuOpen}
-        >
-          <div className="fixed inset-0 z-50" />
-          <Dialog.Panel className="fixed inset-y-0 right-0 z-50 w-full overflow-y-auto bg-gray-900 px-6 py-6 sm:max-w-sm sm:ring-1 sm:ring-white/10">
-            <div className="flex items-center justify-between">
-              <a href="#" className="-m-1.5 p-1.5">
-                <span className="sr-only">DataLine</span>
-                <img className="h-8 w-auto" src={logomd} alt="" />
-              </a>
-              <button
-                type="button"
-                className="-m-2.5 rounded-md p-2.5 text-gray-400"
-                onClick={() => setMobileMenuOpen(false)}
-              >
-                <span className="sr-only">Close menu</span>
-                <XMarkIcon className="h-6 w-6" aria-hidden="true" />
-              </button>
-            </div>
-            <div className="mt-6 flow-root">
-              <div className="-my-6 divide-y divide-gray-500/25">
-                <div className="space-y-2 py-6">
-                  {navigation.map((item) => (
-                    <a
-                      key={item.name}
-                      href={item.href}
-                      className="-mx-3 block rounded-lg px-3 py-2 text-base font-semibold leading-7 text-white hover:bg-gray-800"
-                    >
-                      {item.name}
-                    </a>
-                  ))}
-                </div>
-                {/* <div className="py-6">
-                  <a
-                    href="#"
-                    className="-mx-3 block rounded-lg px-3 py-2.5 text-base font-semibold leading-7 text-white hover:bg-gray-800"
-                  >
-                    Log in
-                  </a>
-                </div> */}
-              </div>
-            </div>
-          </Dialog.Panel>
-        </Dialog>
-      </header>
+      <Header></Header>
 
       <div className="relative isolate  pt-12">
         <BackgroundLayout></BackgroundLayout>
@@ -237,7 +128,7 @@ export const Landing = () => {
                 </div>
               </div>
               <img
-                src={connection_creator_image}
+                src={data_security_image}
                 alt="Connection creator screenshot"
                 className="w-[40rem] max-w-none rounded-xl shadow-xl ring-1 ring-white/10 sm:w-[57rem] md:-ml-4 lg:-ml-0"
                 width={2432}
